code/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
Brotli==1.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
dnspython==2.7.0
docker==7.1.0
frozenlist==1.5.0
g4f==0.5.1.1
idna==3.10
iniconfig==2.1.0
motor==3.7.0
multidict==6.4.3
nest-asyncio==1.6.0
packaging==24.2
pluggy==1.5.0
propcache==0.3.1
pycryptodome==3.22.0
pymongo==4.12.0
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
requests==2.32.3
testcontainers==4.10.0
typing_extensions==4.13.2
urllib3==2.4.0
wrapt==1.17.2
yarl==1.20.0
